--- packages/app/src/app/h/[serverId]/schedules.tsx
@@ -1,15 +1,0 @@
-import { Redirect } from "expo-router";
-import { HostRouteBootstrapBoundary } from "@/components/host-route-bootstrap-boundary";
-import { buildSchedulesRoute } from "@/utils/host-routes";
-
-export default function HostSchedulesRoute() {
-  return (
-    <HostRouteBootstrapBoundary>
-      <HostSchedulesRouteContent />
-    </HostRouteBootstrapBoundary>
-  );
-}
-
-function HostSchedulesRouteContent() {
-  return <Redirect href={buildSchedulesRoute()} />;
-}
