+++ packages/app/src/utils/host-routes.ts
@@ -411,14 +411,6 @@ export function buildHostSessionsRoute(serverId: string) {
   return `${base}/sessions` as const;
 }
 
-export function buildHostSchedulesRoute(serverId: string) {
-  const base = buildHostRootRoute(serverId);
-  if (base === "/") {
-    return "/" as const;
-  }
-  return `${base}/schedules` as const;
-}
-
 export function buildSessionsRoute() {
   return "/sessions" as const;
 }
