+++ packages/server/src/server/schedule/cron.ts
@@ -1,101 +1,5 @@
+import { parseCronExpression } from "@getpaseo/protocol/schedule/cron-expression";
 import type { ScheduleCadence } from "@getpaseo/protocol/schedule/types";
-
-interface CronFieldMatcher {
-  matches(value: number): boolean;
-}
-
-function buildValueSet(values: Iterable<number>): Set<number> {
-  return new Set(values);
-}
-
-function createRange(start: number, end: number, step: number): number[] {
-  const values: number[] = [];
-  for (let value = start; value <= end; value += step) {
-    values.push(value);
-  }
-  return values;
-}
-
-function parseStepPart(part: string, bounds: { name: string }): { base: string; step: number } {
-  const stepParts = part.split("/");
-  if (stepParts.length > 2) {
-    throw new Error(`Invalid cron ${bounds.name} step`);
-  }
-
-  const [base, stepSource] = stepParts;
-  const normalizedStep = stepSource?.trim();
-  const step = normalizedStep === undefined ? 1 : Number.parseInt(normalizedStep, 10);
-  if (
-    !Number.isInteger(step) ||
-    step <= 0 ||
-    (normalizedStep !== undefined && String(step) !== normalizedStep)
-  ) {
-    throw new Error(`Invalid cron ${bounds.name} step`);
-  }
-
-  return { base, step };
-}
-
-function parseField(
-  source: string,
-  bounds: { min: number; max: number; name: string },
-): CronFieldMatcher {
-  const trimmed = source.trim();
-  if (!trimmed) {
-    throw new Error(`Invalid cron ${bounds.name} field`);
-  }
-
-  const values = new Set<number>();
-  for (const rawPart of trimmed.split(",")) {
-    const part = rawPart.trim();
-    if (!part) {
-      throw new Error(`Invalid cron ${bounds.name} field`);
-    }
-
-    const { base, step } = parseStepPart(part, bounds);
-
-    if (base === "*") {
-      for (const value of createRange(bounds.min, bounds.max, step)) {
-        values.add(value);
-      }
-      continue;
-    }
-
-    const rangeMatch = base.match(/^(\d+)-(\d+)$/);
-    if (rangeMatch) {
-      const start = Number.parseInt(rangeMatch[1], 10);
-      const end = Number.parseInt(rangeMatch[2], 10);
-      if (start > end || start < bounds.min || end > bounds.max) {
-        throw new Error(`Invalid cron ${bounds.name} range`);
-      }
-      for (const value of createRange(start, end, step)) {
-        values.add(value);
-      }
-      continue;
-    }
-
-    const value = Number.parseInt(base, 10);
-    if (!Number.isInteger(value) || value < bounds.min || value > bounds.max) {
-      throw new Error(`Invalid cron ${bounds.name} value`);
-    }
-    values.add(value);
-  }
-
-  const allowed = buildValueSet(values);
-  return {
-    matches(value: number): boolean {
-      return allowed.has(value);
-    },
-  };
-}
-
-interface ParsedCronExpression {
-  minute: CronFieldMatcher;
-  hour: CronFieldMatcher;
-  dayOfMonth: CronFieldMatcher;
-  month: CronFieldMatcher;
-  dayOfWeek: CronFieldMatcher;
-}
 
 interface CronDateParts {
   minute: number;
@@ -103,21 +7,6 @@ interface CronDateParts {
   dayOfMonth: number;
   month: number;
   dayOfWeek: number;
-}
-
-function parseCronExpression(expression: string): ParsedCronExpression {
-  const parts = expression.trim().split(/\s+/);
-  if (parts.length !== 5) {
-    throw new Error("Cron expressions must have 5 fields");
-  }
-
-  return {
-    minute: parseField(parts[0], { min: 0, max: 59, name: "minute" }),
-    hour: parseField(parts[1], { min: 0, max: 23, name: "hour" }),
-    dayOfMonth: parseField(parts[2], { min: 1, max: 31, name: "day-of-month" }),
-    month: parseField(parts[3], { min: 1, max: 12, name: "month" }),
-    dayOfWeek: parseField(parts[4], { min: 0, max: 6, name: "day-of-week" }),
-  };
 }
 
 function startOfNextMinute(date: Date): Date {
